requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
asttokens==2.4.1
attrs==25.4.0
base58==1.0.3
bitarray==3.8.0
cached-property==2.0.1
cchecksum==0.3.5
certifi==2025.11.12
charset-normalizer==3.4.4
ckzg==2.1.5
click==8.1.8
cytoolz==1.1.0
decorator==5.2.1
eip712==0.3.3
eth-abi==5.2.0
eth-account==0.13.7
eth-ape==0.8.45
eth-bloom==3.1.0
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-pydantic-types==0.2.5
eth-rlp==2.2.0
eth-tester==0.13.0b1
eth-typing==5.2.1
eth-utils==5.3.1
ethpm-types==0.6.29
evm-trace==0.2.6
evmchains==0.1.6
executing==2.2.1
frozenlist==1.8.0
hexbytes==1.3.1
idna==3.11
ijson==3.4.0.post0
iniconfig==2.3.0
ipython==8.37.0
jedi==0.19.2
lark==1.2.2
lazyasd==0.1.4
lru-dict==1.4.1
mamushi==0.0.6
markdown-it-py==4.0.0
matplotlib-inline==0.2.1
mdurl==0.1.2
morphys==1.0
msgspec==0.20.0
multidict==6.7.0
mypy-extensions==1.1.0
numpy==1.26.4
packaging==23.2
pandas==2.3.3
parsimonious==0.10.0
parso==0.8.5
pathspec==0.12.1
pexpect==4.9.0
pluggy==1.6.0
prompt-toolkit==3.0.52
propcache==0.4.1
ptyprocess==0.7.0
pure-eval==0.2.3
py-cid==0.3.1
py-ecc==8.0.0
py-evm==0.12.1b1
py-geth==6.4.0
py-multibase==1.0.3
py-multicodec==0.2.1
pycryptodome==3.23.0
pydantic==2.12.5
pydantic-core==2.41.5
pydantic-settings==2.12.0
pygments==2.19.2
pymultihash==0.8.2
pytest==8.4.2
python-baseconv==1.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
pyunormalize==17.0.0
pyyaml==6.0.3
regex==2025.11.3
requests==2.32.5
rich==13.9.4
rlp==4.1.0
ruff==0.12.8
safe-pysha3==1.0.5
semantic-version==2.10.0
six==1.17.0
sortedcontainers==2.4.0
sqlalchemy==2.0.45
stack-data==0.6.3
toolz==1.1.0
tqdm==4.67.1
traitlets==5.14.3
trie==3.1.0
types-requests==2.32.4.20250913
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.3
urllib3==2.6.2
varint==1.0.2
watchdog==3.0.0
wcwidth==0.2.14
web3==7.14.0
websockets==15.0.1
yarl==1.22.0
